Reorganize files in repo
Other than editing paths in .gitignore, no files were changed, they were just moved. Considering splitting API scripts to their own repo, but just isolating by file organization for now.
