+++ configuration_template.py
